Detect "stale" background tasks.
I saw in development an occurrence where the app was launched in the
background, and an existing old background task was set. This prevented
me from pulling down new messages, until I launched the app in the
foreground.

I haven't reproduced this yet, but I added an assert to detect if it
happens again.

// FREEBIE